add exists check to make links

+++ make_links.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 mkdir -p ~/bin
-for prog in bell
+for prog in bell sec2elap whatsup
 do
-    ln -sv $HOME/bash/$prog.sh $HOME/bin/$prog
+    if [ ! -f $HOME/bin/$prog ]; then
+	ln -sv $HOME/bash/$prog.sh $HOME/bin/$prog
+    fi
 done
